+++ .eslintrc.json
@@ -21,7 +21,7 @@
     },
     "plugins": ["react", "@typescript-eslint"],
     "rules": {
-      // This needs to be fixed and turned back on:
+      // [Issue #74] This needs to be fixed and turned back on:
       "@typescript-eslint/no-explicit-any": "off"
     }
   }
