+++ src/utils/SecurityUtils.tsx
@@ -7,8 +7,8 @@ import CryptoJS from "crypto-js";
 /**
  * Uses AES to encrypt some text with the given private key.
  */
-export function encrypt(normalText: string, privateKey: string): CryptoJS.lib.CipherParams {
-  return CryptoJS.AES.encrypt(normalText, privateKey);
+export function encrypt(normalText: string, privateKey: string): string {
+  return CryptoJS.AES.encrypt(normalText, privateKey).toString(CryptoJS.format.OpenSSL);
 }
 
 /**
